// src/controllers/tasks/schema.ts
import { z } from 'zod';

export const createTaskSchema = z.object({
    title: z.string().min(1, 'Title is required'),
    description: z.string().optional(),
    status: z.enum(['Pending', 'In Progress', 'Done']).optional(),
    assignedTo: z.string().nullable().optional()
});

export const updateTaskSchema = z.object({
    title: z.string().min(1, 'Title is required').optional(),
    description: z.string().optional(),
    status: z.enum(['Pending', 'In Progress', 'Done']).optional(),
    assignedTo: z.string().nullable().optional()
});